fix(5-2): fix store
