+++ src/store/cities/cities.test.js
@@ -1,14 +1,13 @@
 import {ActionCreator, ActionType} from "./actions.js";
 import {reducer} from "./reducer.js";
 
-import {CITIES, SortType} from "../../utils/constants.js";
+import {CITIES} from "../../utils/constants.js";
 
 describe(`Reducer work correctly`, () => {
   it(`Reducer without additional parameters should return initial state`, () => {
     expect(reducer(undefined, {})).toEqual({
       cities: CITIES,
       currentCity: CITIES[0],
-      currentSort: SortType.DEFAULT,
     });
   });
 
@@ -16,31 +15,15 @@ describe(`Reducer work correctly`, () => {
     expect(reducer({
       cities: CITIES,
       currentCity: CITIES[0],
-      currentSort: SortType.DEFAULT,
     }, {
       type: ActionType.SET_CURRENT_CITY,
       payload: CITIES[1],
     })).toEqual({
       cities: CITIES,
       currentCity: CITIES[1],
-      currentSort: SortType.DEFAULT,
     });
   });
 
-  it(`Reducer should set current sort by a given value`, () => {
-    expect(reducer({
-      cities: CITIES,
-      currentCity: CITIES[0],
-      currentSort: SortType.DEFAULT,
-    }, {
-      type: ActionType.SET_CURRENT_SORT,
-      payload: SortType.TOP_RATED,
-    })).toEqual({
-      cities: CITIES,
-      currentCity: CITIES[0],
-      currentSort: SortType.TOP_RATED,
-    });
-  });
 });
 
 describe(`Action creators work correctly`, () => {
@@ -48,13 +31,6 @@ describe(`Action creators work correctly`, () => {
     expect(ActionCreator.setCurrentCity(CITIES[1])).toEqual({
       type: ActionType.SET_CURRENT_CITY,
       payload: CITIES[1],
-    });
-  });
-
-  it(`Action creator for setCurrentSort returns correct action`, () => {
-    expect(ActionCreator.setCurrentSort(SortType.TOP_RATED)).toEqual({
-      type: ActionType.SET_CURRENT_SORT,
-      payload: SortType.TOP_RATED,
     });
   });
 });
