+++ src/store/offers/actions.js
@@ -1,6 +1,7 @@
 export const ActionType = {
   LOAD_OFFERS: `LOAD_OFFERS`,
   SET_CURRENT_OFFER: `SET_CURRENT_OFFER`,
+  SET_CURRENT_SORT: `SET_CURRENT_SORT`,
 };
 
 export const ActionCreator = {
@@ -13,4 +14,10 @@ export const ActionCreator = {
     type: ActionType.SET_CURRENT_OFFER,
     payload: offer
   }),
+
+  setCurrentSort: (sort) => ({
+    type: ActionType.SET_CURRENT_SORT,
+    payload: sort,
+  }),
+
 };
