+++ src/store/offers/reducer.js
@@ -1,9 +1,11 @@
 import {extendObject} from "../../utils/utils.js";
 import {ActionType} from "./actions.js";
+import {SortType} from "../../utils/constants.js";
 
 const initialState = {
   offers: [],
   currentOffer: null,
+  currentSort: SortType.DEFAULT,
 };
 
 export const reducer = (state = initialState, action) => {
@@ -16,6 +18,11 @@ export const reducer = (state = initialState, action) => {
     case ActionType.SET_CURRENT_OFFER:
       return extendObject(state, {
         currentOffer: action.payload,
+      });
+
+      case ActionType.SET_CURRENT_SORT:
+      return extendObject(state, {
+        currentSort: action.payload,
       });
 
     default: return state;
